+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to `mixislv/laravel-mandrill-driver` will be documented in this file
 
+## 1.1.0 - 2019-11-01
+- Set minimum Laravel version to 6.0
+- Added Mandrill webhook's support
 
 ## 1.0.1 - 2019-09-30
 - X-Mandrill-* to X-MC-*